requirements.txt
PyMySQL
flask
flask_login
cryptography
pyyaml
GitHub-Flask
Flask-SQLAlchemy
pymongo
numpy
gunicorn

